+++ execution/execmodule/chainreader/chain_reader.go
@@ -370,10 +370,10 @@ func (c ChainReaderWriterEth1) GetAssembledBlock(ctx context.Context, id uint64)
 	withdrawals := solid.NewStaticListSSZ[*cltypes.Withdrawal](int(clparams.MainnetBeaconConfig.MaxWithdrawalsPerPayload), 44)
 	for _, w := range block.Withdrawals() {
 		withdrawals.Append(&cltypes.Withdrawal{
-			Amount:    w.Amount,
+			Amount:    uint64(w.Amount),
 			Address:   w.Address,
-			Index:     w.Index,
-			Validator: w.Validator,
+			Index:     uint64(w.Index),
+			Validator: uint64(w.Validator),
 		})
 	}
 	eth1Block.Withdrawals = withdrawals
